+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
-# 1.2
+# 0.1.2
+* fix indexing bug in VBox
+* add interactivity to VBox demo application
+* minor improvements to event bus system
+* improve appearance of focusable widgets
 
-# 1.1
+# 0.1.1
 * project has a demos/ directory
 * ListBox scrolling implemented
 * integration of UNIX signal handlers